Collect Attribute and Event in events.rs
