+++ packages/std/src/results/submessages.rs
@@ -4,7 +4,7 @@ use std::fmt;
 
 use crate::{Binary, ContractResult};
 
-use super::{Attribute, CosmosMsg, Empty};
+use super::{CosmosMsg, Empty, Event};
 
 /// Use this to define when the contract gets a response callback.
 /// If you only need it for errors or success you can select just those in order
@@ -115,55 +115,4 @@ pub struct Reply {
 pub struct SubMsgExecutionResponse {
     pub events: Vec<Event>,
     pub data: Option<Binary>,
-}
-
-/// A full Cosmos SDK event as documented in
-/// https://docs.cosmos.network/v0.42/core/events.html.
-///
-/// This version uses string attributes (similar to
-/// https://github.com/cosmos/cosmos-sdk/blob/v0.42.5/proto/cosmos/base/abci/v1beta1/abci.proto#L56-L70),
-/// which then get magically converted to bytes for Tendermint somewhere between
-/// the Rust-Go interface, JSON deserialization and the `NewEvent` call in Cosmos SDK.
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct Event {
-    /// The event type. This is renamed to "ty" because "type" is reserved in Rust. This sucks, we know.
-    #[serde(rename = "type")]
-    pub ty: String,
-    pub attributes: Vec<Attribute>,
-}
-
-impl Event {
-    /// Create a new event with the given type and an empty list of attributes.
-    pub fn new(ty: impl Into<String>) -> Self {
-        Event {
-            ty: ty.into(),
-            attributes: Vec::with_capacity(10),
-        }
-    }
-
-    /// Add an attribute to the event.
-    pub fn attr(mut self, key: impl Into<String>, value: impl Into<String>) -> Self {
-        self.attributes.push(Attribute {
-            key: key.into(),
-            value: value.into(),
-        });
-        self
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::attr;
-
-    #[test]
-    fn event_construction() {
-        let event_direct = Event {
-            ty: "test".to_string(),
-            attributes: vec![attr("foo", "bar"), attr("bar", "baz")],
-        };
-        let event_builder = Event::new("test").attr("foo", "bar").attr("bar", "baz");
-
-        assert_eq!(event_direct, event_builder);
-    }
 }
